moondev/requirements.txt
anthropic>=0.40.0
openai>=1.59.5
groq>=0.16.0
backtesting>=0.3.3
pandas-ta>=0.3.14b0
yfinance>=0.2.40
ccxt>=4.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
rich>=13.0.0
